+++ src/TransitionPacket.cs
@@ -10,5 +10,6 @@ namespace Dumbwaiter
         [ProtoMember(3)] public int SourceY;
         [ProtoMember(4)] public int SourceZ;
         [ProtoMember(5)] public float Volume;
+        [ProtoMember(6)] public bool SuppressFade; // cargo send: sounds only, no blackout
     }
 }